Removed Api client from Web.Server
